requirements/prod.txt
Django==5.0.4
django-environ==0.11.2
pytz==2024.1
Pillow==10.3.0
celery==5.4.0
redis==5.0.4
cryptography==42.0.5
djangorestframework==3.15.1
dj-rest-auth==6.0.0
